--- app/constants/AppStyles.ts
@@ -1,8 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-module.exports = StyleSheet.create({
-  text: {
-    marginBottom: 10,
-    textAlign: 'center'
-  }
-});
