--- app/constants/colours.less
@@ -1,25 +1,0 @@
-// LIGHT
-@background: #fff;
-@navBackground: #fff;
-@navBorder: #e3e3e3;
-@tabIconActive: #000;
-@tabIconDefault: #8a8a8a;
-@text: #000;
-@icon: #000;
-@seperator: #bdc3c7;
-@monthItem: #e6e6f9;
-@dayItem: #e5edf9;
-@btn: #e6e6f9;
-
-// DARK
-@background: #2c2c54;
-@navBackground: #131324;
-@navBorder: #e3e3e3;
-@tabIconActive: #fff;
-@tabIconDefault: #8a8a8a;
-@text: #fff;
-@icon: #fff;
-@seperator: #4c4c75;
-@monthItem: #131324;
-@dayItem: #171b43;
-@btn: #131324;
