+++ app/constants/styles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+const globalStyles = StyleSheet.create({
+  divider: {
+    alignSelf: 'center',
+    marginVertical: 20,
+    height: 1,
+    width: '70%'
+  }
+});
+
+export default globalStyles;
